Move DateTimeFormat to naviita common
