+++ source/jormungandr/jormungandr/interfaces/v1/LineReports.py
@@ -31,10 +31,10 @@
 
 from __future__ import absolute_import, print_function, unicode_literals, division
 
-from navitiacommon.parser_args_type import BooleanType
+from navitiacommon.parser_args_type import BooleanType, DateTimeFormat
 from jormungandr import i_manager, timezone
 from jormungandr.interfaces.argument import ArgumentDoc
-from jormungandr.interfaces.parsers import DateTimeFormat, default_count_arg_type
+from jormungandr.interfaces.parsers import default_count_arg_type
 from jormungandr.interfaces.v1.decorators import get_obj_serializer
 from jormungandr.interfaces.v1.errors import ManageError
 from jormungandr.interfaces.v1.fields import (
@@ -53,7 +53,6 @@ from jormungandr.interfaces.v1.ResourceUri import ResourceUri
 from jormungandr.interfaces.v1.serializer import api
 from jormungandr.resources_utils import ResourceUtc
 from jormungandr.utils import date_to_timestamp
-
 from flask_restful import fields, reqparse
 from flask.globals import g
 from datetime import datetime
